refactor: team 패키지에 존재하는 전시 관련 dto를 teamdisplay 패키지로 옮기기
