+++ backend/src/main/java/com/woowacourse/levellog/teamdisplay/dto/ParticipantDto.java
@@ -1,4 +1,4 @@
-package com.woowacourse.levellog.team.dto;
+package com.woowacourse.levellog.teamdisplay.dto;
 
 import com.woowacourse.levellog.team.domain.Participant;
 import lombok.AccessLevel;
