+++ backend/src/main/java/com/woowacourse/levellog/teamdisplay/dto/TeamListDto.java
@@ -1,4 +1,4 @@
-package com.woowacourse.levellog.team.dto;
+package com.woowacourse.levellog.teamdisplay.dto;
 
 import java.util.List;
 import lombok.AccessLevel;
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @EqualsAndHashCode
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
-public class TeamsDto {
+public class TeamListDto {
 
     private List<TeamDto> teams;
 }
